requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pymongo==4.4.0
requests==2.31.0
Shapely==2.1.1
geopy==2.4.0